+++ src/Plugins/Tags/App/Taggable/TaggableRepository.php
@@ -4,9 +4,9 @@ namespace Thinktomorrow\Chief\Plugins\Tags\App\Taggable;
 
 interface TaggableRepository
 {
-    public function attachTags(array $taggableIds, array $tagIds): void;
+    public function attachTags(string $taggableType, array $taggableIds, array $tagIds): void;
 
-    public function detachTags(array $taggableIds, array $tagIds): void;
+    public function detachTags(string $taggableType, array $taggableIds, array $tagIds): void;
 
-    public function syncTags(Taggable $taggable, array $tagIds): void;
+    public function syncTags(string $taggableType, array $taggableIds, array $tagIds): void;
 }